# cloudnet-rest-module/src/main/resources/documentation/swagger.yaml
openapi: 3.1.0
info:
  description: >
    CloudNet's rest API version 3 is the next generation in CloudNet's rest
    history.
    
    
    In order to access rest API v3 you will have to authenticate yourself using basic auth.
    To create a basic auth user use the `rest user create <id> <password>` command on your cloudnet node.
    After authenticating with basic auth on `/auth` you will get a bearer jwt access token that must be used for further
    requests to the api. For further information about the response take a look at the `/auth` endpoint.
    
    MENTION CONFIG HERE
    MENTION CASE-SENS HERE
  version: '3'
  title: CloudNet
  contact:
    name: CloudNet
    url: https://cloudnetservice.eu
  license:
    name: Apache license, version 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: '{protocol}://{host}:{port}/api/v3/'
    variables:
      protocol:
        enum:
          - http
          - https
        default: https
      host:
        default: 127.0.0.1
      port:
        default: '2812'
security:
  - bearerAuth: [ ]
...
paths:
...
  /player/onlineCount:
    get:
      tags:
        - Player
      summary: Get the number of online players
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_bridge:player_read`

          - `cloudnet_bridge:player_online_count`
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - properties:
                      onlineCount:
                        type: integer
                        example: 10
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /player/online:
    parameters:
      - name: limit
        in: query
        required: false
        description: The maximum amount of players with which the API responds
        examples:
          limitTen:
            summary: Limit the response to 10 players
            value: 10
        schema:
          type: integer
          minimum: 1
          default: 10
          maximum: 25
      - name: offset
        in: query
        required: false
        description: The amount of players to skip when responding
        examples:
          skipTen:
            summary: Skip the first 10 players
            value: 10
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: sort
        in: query
        required: false
        description: Sorts the players by their name
        examples:
          sortAsc:
            summary: The players will be sorted in ascending order by their name
            value: asc
          sortDesc:
            summary: The players will be sorted in descending order by their name
            value: desc
        schema:
          type: string
          enum:
            - asc
            - desc
          default: asc
    get:
      tags:
        - Player
      summary: Get a list of online players based on the query parameters
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_bridge:player_read`
        
          - `cloudnet_bridge:player_get_bulk`
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CloudPlayer'
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: The total count of players that are online
            X-Result-Count:
              schema:
                type: integer
              description: The count of results returned in this response
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /player/online/{identifier}:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the player to get
        example: bcc582ed-494d-4b93-86cb-b58564651a26
        schema:
          type: string
    get:
      tags:
        - Player
      summary: Get a player by his unique id or name
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_bridge:player_read`

          - `cloudnet_bridge:player_get`
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - properties:
                      player:
                        $ref: '#/components/schemas/CloudPlayer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /player/online/{identifier}/exists:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the player to check for
        example: bcc582ed-494d-4b93-86cb-b58564651a26
        schema:
          type: string
    get:
      tags:
        - Player
      summary: Checks if a specific player is registered
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_bridge:player_read`

          - `cloudnet_bridge:player_exists`
      responses:
        '204':
          description: OK
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /player/online/{identifier}/connectService:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the player to check for
        example: bcc582ed-494d-4b93-86cb-b58564651a26
        schema:
          type: string
      - name: target
        in: query
        required: true
        description: The name of the service to connect the player to.
        schema:
          type: string
        examples:
          serviceConnect:
            summary: Example to connect the player to a service called `SkyWars-1`
            value: SkyWars-1
    post:
      tags:
        - Player
      summary: Connect a specific player to a specific service.
      description: |
        Connects the given player to the specified service.
        
        One of the following scopes is needed to execute the request:
          - `cloudnet_bridge:player_write`
        
          - `cloudnet_bridge:player_connect_service`
      responses:
        '204':
          description: The request to connect the player was sent.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /player/online/{identifier}/connect:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the player to check for
        example: bcc582ed-494d-4b93-86cb-b58564651a26
        schema:
          type: string
      - name: target
        in: query
        required: true
        description: |
          The name of the task / group to connect the player to.
          Use `type` to determine whether the target is a group or a task.
        schema:
          type: string
        examples:
          taskConnect:
            summary: Example to connect the player to a task called `SkyWars`
            value: SkyWars
          groupConnect:
            summary: Example to connect the player to a group called `Fallbacks`
            value: Fallbacks
      - name: serverSelector
        in: query
        required: true
        description: |
          The server selector determines which service from the given task / group to connect to.
        schema:
          type: string
          enum:
            - LOWEST_PLAYERS
            - HIGHEST_PLAYERS
            - RANDOM
        examples:
          lowestPlayers:
            summary: Example to connect the player to the service of the task / group with the lowest players.
            value: LOWEST_PLAYERS
      - name: type
        in: query
        required: true
        description: |
          The type that determines whether the given target is a task or a group.
        schema:
          type: string
          enum:
            - TASK
            - GROUP
        examples:
          groupConnect:
            summary: The given target is a group.
            value: GROUP
    post:
      tags:
        - Player
      summary: Connects a player to a given task or group.
      description: |
        Connects the given player to the given task or group based on the given target & type.

        One of the following scopes is needed to execute the request:
          - `cloudnet_bridge:player_write`
        
          - `cloudnet_bridge:player_connect_group_task`
      responses:
        '204':
          description: The request to connect the player was sent.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /player/online/{identifier}/connectFallback:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the player to check for
        example: bcc582ed-494d-4b93-86cb-b58564651a26
        schema:
          type: string
    post:
      tags:
        - Player
      summary: Connects a given player to a fallback service.
      description: |
        Connects the given player to any fallback service that matches the requirements.
        
        One of the following scopes is needed to execute the request:
          - `cloudnet_bridge:player_write`
        
          - `cloudnet_bridge:player_connect_fallback`
      responses:
        '204':
          description: The request to connect the player was sent.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /player/online/{identifier}/kick:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the player to check for
        example: bcc582ed-494d-4b93-86cb-b58564651a26
        schema:
          type: string
    post:
      tags:
        - Player
      summary: Kicks a given player from the network.
      description: |
        Kicks the given player from the whole network.

        One of the following scopes is needed to execute the request:
          - `cloudnet_bridge:player_write`
        
          - `cloudnet_bridge:player_disconnect`
      requestBody:
        description: The body containg the message to kick the player with.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kickMessage:
                  type: string
      responses:
        '204':
          description: The request to kick the player was sent.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /player/online/{identifier}/sendChat:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the player to check for
        example: bcc582ed-494d-4b93-86cb-b58564651a26
        schema:
          type: string
    post:
      tags:
        - Player
      summary: Sends a chat message to a given player.
      description: |
        Sends the given chat message to the given player.

        One of the following scopes is needed to execute the request:
          - `cloudnet_bridge:player_write`
        
          - `cloudnet_bridge:player_send_chat`
      requestBody:
        description: The body containg the chat message to send to the player.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chatMessage:
                  type: string
      responses:
        '204':
          description: The chat message was sent to the player.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /player/online/{identifier}/command:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the player to check for
        example: bcc582ed-494d-4b93-86cb-b58564651a26
        schema:
          type: string
      - name: redirectToServer
        in: query
        required: false
        description: Whether to redirect the command to the downstream service if the command is not available on the proxy.
        schema:
          type: boolean
    post:
      tags:
        - Player
      summary: Executes a given command for a player.
      description: |
        Executes the given command for the given player.

        One of the following scopes is needed to execute the request:
          - `cloudnet_bridge:player_write`
        
          - `cloudnet_bridge:player_send_command`
      requestBody:
        description: The body containg the command to execute for the player. Its mandatory to **not** have a slash at the start of the command.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
      responses:
        '204':
          description: The command was executed by the player
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    Problem:
      description: The action couldn't be exececuted
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unauthorized:
      description: The the credentials for a successful login are missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: |
        The user has not the required permissions to execute that action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: >
        The request sent to the server does not contain all necessary
        information

        or a specific information provided is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Problem:
      description: >-
        An object describing a problem as specific in RFC 9457. Note that this
        object can contain any number of additional properties depending on the
        called route and error cause.
      type: object
      properties:
        type:
          type: string
          format: uri
          description: An absolute URI that identifies the problem type
          default: about:blank
        title:
          type: string
          description: >-
            A short summary of the problem type. Written in English and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized).
          example: Service Unavailable
        status:
          type: integer
          format: int32
          description: >-
            The HTTP status code generated by the origin server for this
            occurrence of the problem.
          minimum: 400
          exclusiveMaximum: 600
          example: 503
        detail:
          type: string
          description: >-
            A human-readable explanation specific to this occurrence of the
            problem
        instance:
          type: string
          format: uri
          description: >-
            An absolute URI reference that identifies the specific occurrence of
            the problem. It may or may not yield further information if
            dereferenced.
...
    JsonDocPropertyable:
      type: object
      properties:
        properties:
          type: object
          examples:
            - { }
...
    CloudPlayer:
      type: object
      allOf:
        - $ref: '#/components/schemas/CloudOfflinePlayer'
        - properties:
            networkPlayerProxyInfo:
              $ref: '#/components/schemas/NetworkPlayerProxyInfo'
            connectedService:
              $ref: '#/components/schemas/NetworkServiceInfo'
            networkPlayerServerInfo:
              $ref: '#/components/schemas/NetworkPlayerServerInfo'
    CloudOfflinePlayer:
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonDocPropertyable'
        - properties:
            name:
              type: string
              example: derklaro
            firstLoginTimeMillis:
              type: number
              example: 1642409856
            lastLoginTimeMillis:
              type: number
              example: 1642409856
            lastNetworkPlayerProxyInfo:
              $ref: '#/components/schemas/NetworkPlayerProxyInfo'
    NetworkPlayerServerInfo:
      type: object
      properties:
        uniqueId:
          type: string
          example: 05bf52c6-7bb0-4f13-8951-0e1fd803df35
        name:
          type: string
          example: taginbert
        xBoxId:
          type: string
          example: null
        address:
          $ref: '#/components/schemas/HostAndPort'
        networkService:
          $ref: '#/components/schemas/NetworkServiceInfo'
    NetworkPlayerProxyInfo:
      type: object
      allOf:
        - $ref: '#/components/schemas/NetworkPlayerServerInfo'
        - properties:
            version:
              type: integer
              example: 47
            listener:
              $ref: '#/components/schemas/HostAndPort'
            onlineMode:
              type: boolean
              example: true
    NetworkServiceInfo:
      type: object
      properties:
        groups:
          type: array
          items:
            type: string
        serviceId:
          $ref: '#/components/schemas/ServiceId'
...
    ServiceId:
      type: object
      properties:
        taskName:
          type: string
          example: Lobby
        nameSplitter:
          type: string
          example: '-'
        environmentName:
          type: string
          example: MINECRAFT_SERVER
        allowedNodes:
          type: array
          items:
            type: string
          example:
            - Node-1
            - Node-2
        uniqueId:
          type: string
          example: 43a9e824-cc3c-459b-b574-4f41e6509bda
        taskServiceId:
          type: integer
          example: 2
        nodeUniqueId:
          type: string
          example: Node-2
        environment:
          $ref: '#/components/schemas/ServiceEnvironmentType'
...
    HostAndPort:
      type: object
      properties:
        host:
          type: string
          example: 127.0.0.1
        port:
          type: integer
          example: 1410
...
    ServiceEnvironmentType:
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonDocPropertyable'
        - properties:
            name:
              type: string
              example: MINECRAFT_SERVER
            defaultServiceStartPort:
              type: integer
              example: 25565
            defaultProcessArguments:
              type: array
              items:
                type: string
...
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
          description: This boolean will always be false when an error occurred
        reason:
          type: string
          example: Error message
          description: The error that occurred during the processing of the request